+++ src/lib.rs
@@ -13,6 +13,7 @@ pub mod nifs;
 pub mod plonk;
 pub mod polynomial;
 pub mod poseidon;
+pub mod table;
 pub mod util;
 
 pub mod error;
